requirements.txt
pytest
pylint
black
requests
types-requests
pre-commit
pyyaml
types-PyYAML
mypy